requirements.txt
accelerate==1.12.0
anyio==4.12.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
mysqlclient==2.2.7
networkx==3.6.1
num2words==0.5.14
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
service-identity==24.2.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.7.1
torchvision==0.24.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@a7f29523361b2cc12e51c1f5133d95f122f6f45c
Twisted==25.5.0
txaio==25.12.2
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.243
ultralytics-thop==2.0.18
urllib3==2.6.2
websockets==16.0
zope.interface==8.1.1
